displaying reservoir status
