+++ MockKitUI/MockPumpManager+UI.swift
@@ -31,7 +31,7 @@ extension MockPumpManager: PumpManagerUI {
     }
 
     public func settingsViewController(bluetoothProvider: BluetoothProvider, colorPalette: LoopUIColorPalette, allowDebugFeatures: Bool, allowedInsulinTypes: [InsulinType]) -> PumpManagerViewController {
-        let settings = MockPumpManagerSettingsView(pumpManager: self, supportedInsulinTypes: allowedInsulinTypes)
+        let settings = MockPumpManagerSettingsView(pumpManager: self, supportedInsulinTypes: allowedInsulinTypes, appName: appName)
         let hostingController = DismissibleHostingController(rootView: settings, colorPalette: colorPalette)
         hostingController.navigationItem.backButtonDisplayMode = .generic
         let nav = PumpManagerSettingsNavigationViewController(rootViewController: hostingController)
