+++ MockKitUI/ViewModifier/OpenMockPumpSettingsOnLongPressGesture.swift
@@ -49,33 +49,3 @@ fileprivate struct OpenMockPumpSettingsOnLongPressGesture: ViewModifier {
         }
     }
 }
-
-struct MockPumpManagerControlsView: UIViewControllerRepresentable {
-//    @EnvironmentObject var onboardingViewModel: OnboardingViewModel
-//    @Environment(\.dismissAction) var dismiss
-
-    private let pumpManager: MockPumpManager
-    private let supportedInsulinTypes: [InsulinType]
-
-    init(pumpManager: MockPumpManager, supportedInsulinTypes: [InsulinType]) {
-        self.pumpManager = pumpManager
-        self.supportedInsulinTypes = supportedInsulinTypes
-    }
-
-    final class Coordinator: NSObject {
-        private let parent: MockPumpManagerControlsView
-
-        init(_ parent: MockPumpManagerControlsView) {
-            self.parent = parent
-        }
-    }
-
-    func makeUIViewController(context: Context) -> UIViewController {
-        return MockPumpManagerSettingsViewController(pumpManager: pumpManager, supportedInsulinTypes: supportedInsulinTypes)
-    }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
-
-    func makeCoordinator() -> Coordinator { Coordinator(self) }
-}
-
